Move function to proper controller
